Add looping option to settings popup
Now, ideally, we would want that placed in the scroll bar, but this was the occasion to initiate the performer tab... 

I had underestimated how hard this would be to integrate purely do to bad technical choices on my end. The lack of mode unification strikes once again as the various fixtures used to replace it (isModeSilent, toggleListen...) all have to be used to communicate that silent mode is toggled once the loop end is reached. 

Other than that, not much to remark : just a slight redefinition of #getNextIndex().