Adjust velocities of individual channels in settings
This is the first approximation of issue #10. There are sliders allowing for multiplication of velocity from 0 to 2 ; 0 mutes the channel, as the IOCTL is instructed to ignore any events that occur on a channel with a velocity of 0. A helpful toggle is provided alongside each slider to set the gain to 0 or 1. 

Although this works, there are some quirks : 

- The number inputs of the sliders only work with integers, such that the percentages are expressed from 0 to 200, incurring an incoming floor operation that cannot be replicated on the outcoming step. This may lead to float imprecisions. 
- The visualizers will still highlight some (but not all ?) muted channels, and muting them in the middle of the song may leave some notes forever highlighted.
- Most importantly : the muted channels continue to occur as set markers, and so the user still has to perform sets where no notes are actually translated to sound. This one in particular needs fixing.