Introduce toggling of performer constructor options with unmeet
All ChronologyOptions (barring the shift mode which is still finicky and wouldn't see much use for now) are made available in the default settings. They are packaged together even in the store (modifying even one of them means reconstructing the performer anyway, so there is zero reason to separate them). 

For now, only unmeet has been added as a toggle to the settings popup, and it works exactly as intended.