Change velocity percentage gain to offset as requested by Jean Haury
The offset goes from -64 to 64. This means the sliders had to be adapted to support negative values (which they technically did, it's just that their cursor, bar fill and drag did not follow). 

Because no offset equals "channel muted", this information was moved to a separate array of boolean flags. The two are always packaged together, even in the store.

The IOManager also updates this information as it loads, which it previously did not. 

Finally, and also per J. Haury's request, only channels present in the file are displayed. Those that are muted are transparent.