Add chord velocity mapping strategy to settings
MFP.js now carries the information of the "preferred" strategy (the one that has been chosen in settings) and of the *current* strategy (because, as we remember listen mode with no registered perform velocity must always have a strategy of None). It uses a wrapper that updates the inner MFP and sets the current strategy.