First (incomplete) version of grace note support
Grace notes are a very complex and self-contradictory part of musical
notation that translate into a complex MusicXML entity. They can change
the durations of notes that follow and/or precede them, or be inserted
within a measure with a duration of their own, independent of the
measure's own structure.

Due to this complexity, grace notes are buggy in a great majority of
software that handles MusicXML. Among them, OSMD does not yet support
grace notes as valid cursor entries ; thus, any MusicXML file containing
grace notes currently cause a cursor bijection failure leading to cursor
desync. This will only be patched by ignoring grace note in noteSequence
bijection lookup.

At the parse level, the current approach is insufficient for multiple
reasons :
- it does not take after-graces into account (these are grace notes that
follow a note, and thus make use of steal-time-previous, implicitly so
if they are found at the end of a measure)
- even if it did, recomputing the duration of a previous event is not
yet done due to needing to modify the produced MIDI note off event —
although the map to do so is in place — and then rewind the delta
accumulator to account for it
- it considers all default grace note durations to be the same,
regardless of the grace note's notated value (which is one possible
convention, but which might be undesirable for some particular scores)
- it throws an error if the make-time attribute is set, as no test suite
exists which includes it, and as such, its role and implementation
remains unclear even after extensive research and discussion with OSMD
developers.