Support after-graces and mixed after/normal grace clusters
This addresses the first and second bullet points of the second commit.
Note that durations still do not take into account the notated value of
the grace note.

As for make-time, instead of errorring out, a warning is displayed and
notes that may use the attribute will be treated like ordinary grace
notes for now.