NOT WORKING : Tried fixing OSMD cursor in case of grace notes
The OSMD cursor desyncs in certain cases where grace notes are 
encountered. This commit had to be made because, as my mission ends, I 
have no more time to correct this problem ; unfortunately, it is still 
not solved. 

The approach I took to attempt to solve the issue was as follows : 

- Gather all grace notes as noteOns in the MusicXML parser, just like 
tempo and channel changes 
- Attach to each grace note its principal's noteOn
- Convert the deltas to seconds by factorizing the calculation method 
used for the cursor date
- Once the noteSequence is received, replace each grace note and each 
principal by their corresponding object in the sequence
- When setting up the cursor, modify its behavior when grace notes are 
encountered. More specifically : 
- If either A. the current examined set to match to a cursor position 
contains a grace note or B. the next set does, and we are inside a 
cluster (i.e. : a principal is awaited), then search for the position 
starting from *the last position*, not the last position + 1 as was done 
before, to allow for the cursor to occupy the same position multiples 
times. This only happens if the next set does not contain a 
non-principal non-grace note, in which case it must move on regardless 
of the status of the grace note cluster, to meet that note where it is.

This aimed to prevent cursor desync ; attaching notehead listeners would 
have been the next step. But even cursor desync still happens, because 
grace notes happen in intervals too small and may throw off the date 
approximation.

J. Larralde and I, while discussing this issue, realized that my parsing 
of grace notes is probably wrong, in that they should not be allowed to 
interlace with non-grace (which might be an oversimplification, but 
would make the algorithm easier) ; however, this would still not fix the 
date approximation problem, which in the end, seems to be the crux of 
the issue.

There has to be a solution to this, but so far, I have not found it. I 
hope to work on this problem again soon and figure it out.