Fix OSMD cursor desync on files containing grace notes
Replace transmission of grace notes with approximated durations, which
led to date precision errors and missing graces, with the simple
transmission of indices for the grace notes and their corresponding
principals.

This is done by gathering all note ons in the MusicXML parser. The order
thus obtained is *roughly* the same as the NoteSequence's ; the
principal may however be off, but at the very least, the index will
always point to a note in the same set.

The result is that the SheetMusic.vue algorithm, which looks to see if
the "principal" is contained within a set, now succeeds all the time. In
other words, the only possible remaining point of failure is the
condition on which the cursor is asked to remain in place. So far, this
condition has proven efficient.

This version has been tested to work on both K279-1 and Sonate
Chisamori. It should be tried more thoroughly on a set of grace
note-bearing files.

The next step is to restore listeners to notes skipped by the cursor as
it moves forwards through grace note clusters.