Fix handling of aftergraces by cursor and simplify s-i-p condition
Now, if the "upcoming" principal (= the latest one) is detected as 
coming before the current set (i.e. : we have a homogenous 
aftergrace-only cluster), this principal is not saved and processing 
continues as usual.

Note that some aftergraces may still be deprived of listeners : see 
#100. 

In addition to this, simplified the stay-in-place condition to simply 
check the existence of an upcoming principal ; because if there are 
grace notes in the set that we care about, then it exists anyway. This 
makes things more readable.