Fix left-behind system to account for multiple adjacents of same key
graceAdjacentsLeftBehind is now a map of sets. Every time a notehead is 
found to not have an equivalent inside a grace cluster, it is added to 
the set — except if it had already been registered before — and is 
removed only when an equivalent is found later, either through normal 
parsing or looking through unadressed notes of a set. 

usedNotesOfSet and buffersWithListeners have also been converted to 
sets.