Working algorithm for grace notes in OSMD (finally !!!)
This time, grace note cluster processing has been completely separated
from the ordinary date approximation process. It now relies on a map of
sets of every note under the cursor at every eligible position.

When a grace note cluster is reached, the notes under the cursor at the
current position ar examined. If they contain all notes found in the
set, the index is returned immediately. If they do not, the index is
incremented and returned with that increment. If the principal is found
in the set, the current index is returned, but the search index is
incremented so the monotonic search can resume.

With this, the only remaining bugs are :

- In relation to a possible OSMD bug (Chopin BI18op68-2 : two
aftergraces which are separate notes share a cursor position, which is
stuck between two measures ; this throws off the whole setup process)
- In relation to a parser bug that seems to happen around measure edges
(Chopin  op08n01, set 63 ; Grieg op12n01, measure 14 ; Mozart K284-2,
measure 38...)
- More importantly for this specific algorithm : sometimes (a single
example has been found, which is K330-3 set 569), a move that should be
forced is not ; but this induces no cursor desync, as the cursor instead
jumps two steps forwards after this, likely due to the precautionary
supplemental increment added for similar cases.

What is in order, then, is to fix the parser and find the cause of that
last problem ; but this version, at least, is LEAGUES more functional
than any previous one.