Add intra-set pitch order to noteSequence and MusicXML parser
This ensures that the indices calculated in the MusicXML parser are 
exactly those of the noteSequence, and in particular, that a grace 
note's indicated principal is always the correct note of the set, fixing 
cases where it was mistaken for another set member, causing issues in 
e.g. K330-3.

Misc : Remove isAllGraceNoteSet() method from SheetMusic.