Fix off-by-one error in markLastArpeggiatedChordNote
Which led to the invisible error of having two registered chords for 
each : one with only the first note in the chord, another with the 
second registered twice. This was functionally equivalent EXCEPT in a 
file where a measure ends in arps, in which case it would crash.

If I had my eyes open, I could've seen it too, because the correct form 
is literally in the gatherGraceNoteData() function right below :)