Fix possible tiny discrepancies in intra-set pitch order
If two pre-relative deltas were close by a tiny amount, but still 
technically distinct, they would be ordered by delta, not pitch, and as 
a result the resulting indices would be wrong. This led to wrong 
principals and notehead breaks (e.g. K331-3 index 562). 

Introduced an epsilon above which pitch order happens. Hopefully it's 
small enough : if not, this could actually worsen the problem as 
non-simultaneous notes would be re-arranged, leading to false positives 
in moveForced and cursor breaks again.