Arpeggio support for SheetMusic visualizer
Through transmission of arpeggioInfo in the MusicXML parser, as one 
would expect ; however, this also gave way to a complete overhaul of 
findNearestCursorDate.

It has become apparent that date approximation was now entirely 
unneeded. By selecting positions containing only eligible notes, we 
ensure that outside of edge cases like arps and graces, the cursor will 
*always* move forward. In other words, the default action is to return 
the current index, and increment it. 

This has been tested to work identically to the previous method, and 
arpeggios are now fully supported.