Fix arp shifts not happening
The "event" terminology in the MusicXML parser is horrible ; the term 
applies to both MIDI events and XML nodes, and because of this, I lost 
track and checked against an XML node property to detect notes to shift, 
resulting in no shifts happening at all.

This was not visible on files like K279-1 or K280-1 ; the divisions of 
these files was simply too small for a shift to make any difference. 
However, K284-3, which is broken in other ways, revealed this bug at 
measure (notated) 70 (measure index 77 due to implicit measure 
numbering).

As well, applying the shift to events directly *at* the arp delta (>=) 
was incorrect : if they are synced with the arp, they must remain so !