Visualizer support for synchronous arpeggios
Arpeggio info now tests to see if the last note is associated with the 
first, because each arpeggio-ending note carries the reference of its 
starting note. This allows the visualizer to properly know when 
overlapping arpeggios end, and to stay in place as long as one of them 
is going on.

Curiously, the case where synchronous arpeggios *have* been found (Clair 
de Lune) does not suffer any particular desync from the parser's lack of 
specific consideration for such a case. This might indicate that issue 
#106 is not warranted after all.