Fix arpeggio shift applying to own non-first arp notes
...which was the cause of this "crooked" feeling : all notes aside from 
the first in the arp were shifted by the arp's own increment, meaning 
they started with a delay and consequently ended too late, when the 
first note was fine.

This is good in the general case, but it BREAKS the (thus accidental) 
non-problematic nature of synced arpeggios, which was dodged by them 
being mistakenly double-shifted ; hence Clair de Lune is broken.