Fix Clair de Lune sync'd arps (but not general case)
As previously mentioned, fixing arp auto-shift broke Clair de Lune, 
which only "worked" thanks to it. 

The crux of it is : 

- Register *multiple* offsets per delta, under the key of their 
arp-starter
- Do not compensate events synced to an arp if they are themselves part 
of one
- Compensate the events which *are* indeed compensated with the *sum* of 
the offsets registered for that delta
- Do not shift events that are part of an arp which is sync'd with the 
shift's owner
- The contentious part : shift the events by the *difference* between 
the arp's offset and the value of the measure accumulator at the event's 
creation.

This last point is flaky because it only works for two staves at best. 
After that, the difference may not mean much anymore. But I need to move 
on, so for now, it will stay that way.