First (buggy) version of mixed play/perform
...also called "auto-playback" or "auto-listen".

The implementation works based on "triggers" calculated based on given 
channels or tempo indications (beat, measure, etc.). The current state 
of code accepts only channel-based triggers ; channel triggers are 
ignored.

This is a much bigger commit than I originally intended, and even so, 
several imperfections and bugs remain present. 

However, before describing them, it is important to outline that this 
commit touches upon a great amount of the inner workings of MFP.js, so 
here is a little bullet list of all of the aspects in which it does.

- The triggering of perform mode on an incoming command has been moved 
fromm App.vue into MFP.js itself. This is because it is now subject to 
logic accessible only to MFP.js itself.
- The chronology, once obtained in the noteEvents callback, is stored 
inside MFP.js, because it is necessary to calculate playback triggers.
- New flags (yes, even more) have been introduced that control how 
auto-playback behaves. The most crucial is the noInterruptFlag ; it 
prevents auto-playback from being interrupted by a command. 
- The next is the noRenderFlag. It is used to prevent timeouts created 
by auto-playback from being put into effect once perform mode has been 
restored. In addition to this, the timeouts are now cleared *inside* 
playNextSet. This prevents any locking of playback into a positive 
timeout creation feedback loop.
- The emptying of the pending end set has been moved inside the 
noteEvents callback, as it must occur before auto-playback triggers in 
order for it not to be switched out at the wrong time.
- In order to prevent skipping sets when resuming perform, an 
onlyAcceptPressed flag disregards any incoming key release when control 
is restored to the user ; otherwise, their release could trigger another 
auto-playback coming immediately afterwards.
- To prevent auto-playback from triggering on the *wrong* release, a 
semaphore is maintained that ensures all pending releases have occurred 
before it begins. 
- Because the semaphore cannot use an active wait loop (else it would 
freeze all reception of releases and never escape its waiting state), it 
instead resumes the mode setting process in the middle, leading to the 
awkward creation of effectiveSetModeListen(), which contains the 
velocity strategy set and the first call to playNextSet.
- The repeatIndex flag is never set when auto-playback is in progress.
- Finally, rendering itself is now conditional. Rendering is not done 
for a key press on a playback trigger boundary if the semaphore is not 
empty, and for a key release during auto-playback (and right after it, 
as described above). 

As previously said, even all of this does not address every challenge 
this feature poses. Here are just the ones I have been able to identify 
thus far : 

- Very often, a user will press just a little too soon during 
auto-listen, and thus trigger nothing. Something similar to the 
InterruptionGuard is required, but in reverse : a grace period that 
pretends to trigger an event that comes *after* the press, rather than 
before.
- *Some* (not all) notes of auto-playback are played too loud, 
disregarding the proper velocity adaptation.
- All notes started by the last keypress before auto-playback are 
immediately unpainted in the visualizer, although they are not turned 
off
- When index jumping in perform mode, auto-playback repeats the last 
perform set before continuing normally. 
- As perform continues, delay is accumulated that prevents the smooth 
performance of fast rhythms.

Future commits will seek to address these issues.