Grace period before autoswitch to playback for smoother experience
This consists in a flag signaling to store keydowns during the last note 
on of auto-playback, and triggering the last such command in setMode(), 
with the caveat that it will not update the semaphore (doing so would be 
fine most of the time, but causes a rare race condition where the 
semaphore can never again become zero).  

This considerably smooths the playing experience on, e.g., MaouDamashii, 
however fast rhythms (Proof of Courage Chan 5 for instance) remain very 
sluggish.