Move NoteSequence calculation to dedicated standalone util function
This function is called in MFP.js and its result is emitted instead of 
the wholesale chronology. MFP.vue then sets the store state with NS, 
setStarts and setEnds ; and the piano roll simply reads from the store, 
just like SheetMusic does.

This is a necessary prerequisite to tempo-based playback triggers.