Measure-based auto-playback
Aside from channel-based triggers, tempo-based triggers are now 
available, although only for MusicXML files (as measures, beats, etc. 
have no intrinsic meaning in MIDI files). The only variation of them 
implemented here is measure-based : the user must only give one input 
per start of measure. 

The way this is accomplished is by marking noteOns in the MusicXML 
parser. Their indices in the noteSequence (which is now accessible 
directly in loadMidifile()) are used to in turn mark sets that occur at 
the start of measures, and playback triggers are designated as the 
inverse of these.

Misc : move getSet and getSetIndex logic to noteSequenceUtils ; the 
store calls these intermediate versions, the component calls the store's 
wrappers, and MFP.js uses the utils directly.