Toggle for measure-based play
A dedicated component, called PerformGranularity, has been created, 
consisting of an OptionTabs and an illustration.

When a granularity is set, all channels are perform and channel 
selection is disabled. This in turn triggers a flag in MFP.vue not to 
take channel control updates into account.

Finally, when a MIDI file follows a MusicXML file, granularity is 
discarded ; however, it is preserved between MusicXML files.