First version of beat play
For double, triple and quadruple meter, beat boundaries are the 
multiples of the bar subdivision. For compound triple meter, they are 
groups of three subdivisions. 

This is done in spite of delta shifting thanks to a track-spanning 
accumulator of the offset between the "real" delta and the incurred 
delta due to fermata and arpeggios, which is identical to the 
measure-wise offset, except never reset.

However, this does *not* work on files like K279-1, where grace notes in 
non-divisible quantities break up their principal's duration such that 
the delta accumulator never becomes whole again. 

More bugs may also exist.