New algorithm for managing offsets in isOnBeat()
totalOffsets are now only incremented when measures end, with the
*maximum* of their respective offsets, which are themselves updated on
each backup (and on measure ends too).

If an offset has been accrued since the last backup, only it is
considered, and the rest of the measure offset is ignored.

Finally, a backup is considered as "having just performed" during a
whole arpeggio if said arpeggio began right after that backup took
effect.