Fix : add escape hatch for findNearestCursorDate if overshooting positions
This only happens in files where note setup is broken and the cursor was
moved forward in an incorrect place. It prevents setup from hanging and
instead returns early with an incorrect index so that the score can
still be displayed, even if the cursor will break at some point and
hang on the last set.