Account for MusicXML files without measure and/or time signatures
A file is without measures if all of its timewise measures are set to 
implicit. A file without time signatures is without beats.

Under these conditions, measureEnd and measureStarts, as well as 
isOnBeat() are irrelevant. MFP.js picks up on this and sends a signal to 
MFP.vue to deactivate the granularity picker, as well as to OSMD to hide 
the time signatures (because OSMD adds a 4/4 time signature by default).

This has been tested to work with Satie's Gnossienne n°1, with the 
exception of its identically-numbered implicit measures being renumbered 
to distinct identifiers (a bug related to musicxml-interfaces, see 
#142).