First version of autoplay coloring
For SheetMusic cursor and notes, as well as piano roll notes, but not
for the keyboard yet.

I don't really like the chosen shades of yellow, but I'm struggling to
find a color which both contrasts with our chosen green and blue, *and*
looks good besides the two of them.

Misc :

- Fix MFP.js emitting index too early on listen mode (how was this never
caught before ??)