Fix multi-track files with non-default tempo reverting to it
Because the default tempo event present on tracks after the first one 
would cancel its actual setting of the tempo.

Because of this, SheetMusic.vue is not always guaranteed to understand a 
tempo event before the date of position, hence it now produces its own 
default value when such an event is not found.