Fix first notes of MusicXMLs beginning with rests lasting too long
This was a bug introduced by the previous commit, because the relative 
conversion algorithm posited that the first event of every track would 
always be at delta 0 — because it was the default tempo event.

With it gone, this adjustment restores the correct length of first 
non-rest notes in files like Clair de Lune or Parlez-moi.