Auto-playback without necessary key release
The core trick is this : when a playback set is reached, and on keydown, 
two actions are scheduled : 

1. A call to new function getAllPendingNoteOffs() (provided in the new 
MFP lib version), which summons the Renderer's combine3 map to retrieve 
every off event contained therein associated with a currently pressed 
key, thereby only terminating notes which would be turned off in a 
"normal" performance, and leaving notes whose ending lies in future 
starting sets playing as intended. 
2. A call to setMode(), which begins listen mode from the next on set as 
it always does.

The first action is undertaken when the last end set of perform mode 
should occur ; the second, at the time of the first starting set of 
listen mode.

This solves multiple problems : 

- Counting releases is no longer necessary (except maybe for the border 
when still in perform mode, hence the semaphore still exists and remains 
checked just in case).
- Releases can be disabled just as well as keypresses during 
auto-playback, because all of them are now irrelevant since all off 
events linked to a release are triggered when autoplay begins.
- The lack of possible render() calls from key events during auto 
playback negates the possibility of cancelling a render callback, which 
did indeed happen at random times during testing.

This drastically improves the experience of auto-playback.