Minor : cleanup old unused code
In index.js and IOManager, but most importantly by deleting 
NaiveMidifilePerformer, as well as in App.vue, which required a little 
event renaming in the ioctl (because events are apparently not 
segregated based on their emitters !). 

All event handlers registered in App.vue are now properly removed. Of 
course, in practice, this has no effect, but it's good for form. The 
removal of all logic concerning the naive midifile performer also makes 
it all more readable.