Preemptively set noInterrupt flag before scheduling auto listen timeouts
Otherwise, in case of button mashing, the app risked entering an invalid 
state where button presses have an effect in auto-listen, and worse, 
could lock it into auto-listen until forced silent.

This had never been noticed because the call to unfillActiveRects in the 
PianoRoll bought enough time to buffer this possibility, which was made 
inactive in commit b2c2cd7.