Fix files sometimes crashing between loads
Due to the order of watchers and calls, an update to the piano roll's 
base rect color is triggered before it receives a new noteSequence, 
rendering the process invalid. Thankfully, this is overwritten by the 
call to updateBaseRectColor *on* file load, but a dummy value must be 
returned inbetween these.