Try time-based command store instead of fixed last-set storage
To determine if a command should be stored while in autoplay, we check 
if we are close below a threshold to the scheduled end timestamp of the 
autoplay sequence.

Although rather obtuse, the calculation does not appear too time 
consuming. However, it unfortunately seems to have no beneficial effects 
on reactivity either ; files like Proof of Courage remain much slower in 
autoplay vs normal playback. There is a bottleneck somewhere else, but I 
have failed to identify it.